feat: Upgrade JS SDK to `9.16.1` and add durable objects instrumentation (#165)
Durable objects instrumentation support was added in
[9.16.0](https://github.com/getsentry/sentry-javascript/releases/tag/9.16.0)
of the JS SDK. This PR upgrades `sentry-mcp` to get instrumented with
durable objects.

Docs PR for durable objects is here:
https://github.com/getsentry/sentry-docs/pull/13626.

In `packages/mcp-cloudflare/src/server/lib/mcp-transport.ts` I had to
comment out the `wrapMcpServerWithSentry` usage, because this breaks
locally with miniflare. We're following up with the cloudflare team
about this (more details here in linear:
https://linear.app/getsentry/issue/JS-398/9150-breaks-cloudflare-sdk-usage-in-sentry-mcp#comment-7ce26edb).
When we fix this, we'll come back and re-enable it.

`9.16.0` also added support for logs in the Cloudflare SDK, so I enabled
that, and added the `consoleLoggingIntegration` (integration is optional
while logs is experimental in the SDK) to flush logs from `console.X`
calls to Sentry.

---------

Co-authored-by: David Cramer <dcramer@gmail.com>